pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-rl"
version = "1.0.0"
description = "Offloading Topological Reasoning from LLM Agents in the Security Operations Center"
authors = [
    {name = "Uday Hiwase", email = "uday@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["reinforcement-learning", "graph-neural-networks", "cybersecurity", "soc", "neo4j", "ppo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=12.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "streamlit>=1.30.0",
    "networkx>=3.0",
    "pyvis>=0.3.2",
    "requests>=2.31.0",
    "neo4j>=5.14.1",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.1",
    "torch>=2.0.0",
    "torch_geometric>=2.3.0",
    "ray[rllib]>=2.5.0",
    "gymnasium>=0.28.1",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
streaming = [
    "confluent-kafka>=2.3.0",
]

[project.scripts]
sentinel-rl = "src.main:main"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
